Have Vertical and Horizontal Installments separated
