+++ Installments/example.jsx
@@ -1,46 +1,62 @@
 import React from 'react'
-import InstallmentsComponent from '../Installments'
-import UncontrolledInstallments from '../uncontrolled/Installments'
-import { LIVE_WIDE } from '../Showroom/variationTypes'
+import * as Installments from '../Installments'
+import {LIVE, LIVE_WIDE} from '../Showroom/variationTypes'
 
 const options = [
-  { key: 'installments_3', content: <div>Long one line text in div</div> },
-  { key: 'installments_6', content: [<div key='1'>$64.17/mo.</div>, <div key='2'>array of elements</div>] },
+  { key: 'installments_3', content: 'Long one line text in div' },
+  { key: 'installments_6', content: '$64.17/mo. array of elements' },
   { key: 'installments_12', content: 'Simple text content' },
-  { key: 'installments_24', content: <div><div>Force</div><div>new line</div></div> }
+  { key: 'installments_24', content: 'Force new line' }
 ]
 
 export default {
   title: 'Installments',
 
-  examples: {
-    require: `import Installments from '@klarna/ui/Installments'
-import UncontrolledInstallments from '@klarna/ui/uncontrolled/Installments'`,
-    type: LIVE_WIDE,
+  variations: [
+    {
+      title: 'Horizontal',
+      require: `import * as Installments from '@klarna/ui/Installments'`,
+      type: LIVE_WIDE,
 
-    examples: {
-      Regular: <InstallmentsComponent
-        onChange={(key) => console.log('You selected', key)}
-        name='installments'
-        value='installments_12'
-        options={options}
-      />,
+      examples: {
+        Regular: <Installments.Horizontal
+          onChange={(key) => console.log('You selected', key)}
+          name='installments'
+          value='installments_12'
+          options={options}
+        />,
 
-      Uncontrolled: <UncontrolledInstallments
-        onChange={(key) => console.log('You selected', key)}
-        name='installments2'
-        value='installments_24'
-        focus='installments_6'
-        options={options}
-      />,
+        Focused: <Installments.Horizontal
+          onChange={(key) => console.log('You selected', key)}
+          name='installments3'
+          value='installments_12'
+          focus='installments_6'
+          options={options}
+        />
+      }
+    },
 
-      Focused: <InstallmentsComponent
-        onChange={(key) => console.log('You selected', key)}
-        name='installments3'
-        value='installments_12'
-        focus='installments_6'
-        options={options}
-      />
+    {
+      title: 'Vertical',
+      require: `import * as Installments from '@klarna/ui/Installments'`,
+      type: LIVE,
+
+      examples: {
+        Regular: <Installments.Vertical
+          onChange={(key) => console.log('You selected', key)}
+          name='installments'
+          value='installments_12'
+          options={options}
+        />,
+
+      Focused: <Installments.Vertical
+          onChange={(key) => console.log('You selected', key)}
+          name='installments3'
+          value='installments_12'
+          focus='installments_6'
+          options={options}
+        />
+      }
     }
-  }
+  ]
 }
