+++ Radio/example.jsx
@@ -3,7 +3,7 @@ import Radio from '../Radio'
 import * as Checklist from '../Checklist'
 import * as Paragraph from '../Paragraph'
 import { LIVE_WIDE } from '../Showroom/variationTypes'
-import UncontrolledInstallments from '../uncontrolled/Installments'
+import * as Installments from '../Installments'
 import UncontrolledDropdown from '../uncontrolled/Dropdown'
 import Subtitle from '../Subtitle'
 
@@ -37,7 +37,7 @@ const optionsWithContent = [
     label: 'Sit',
     description: 'Amet et consequetur',
     content: <div>
-      <UncontrolledInstallments
+      <Installments.Horizontal
         name='installments'
         onChange={(key) => console.log('You selected', key)}
         options={[{content: <div>Long one line text in div</div>, key: 'installments_3'}, {content: [<div key='1'>$64.17/mo.</div>, <div key='2'>array of elements</div>], key: 'installments_6'}]}
