--- uncontrolled/Installments.jsx
@@ -1,10 +1,0 @@
-import compose from '../lib/compose'
-import statefulValue from '../lib/decorators/statefulValue'
-import statefulOptionFocus from '../lib/decorators/statefulOptionFocus'
-import Installments from '../Installments'
-
-const UncontrolledInstallments = compose(statefulOptionFocus, statefulValue)(Installments)
-
-UncontrolledInstallments.displayName = 'UncontrolledInstallments'
-
-export default UncontrolledInstallments
